back/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "RafaelBochi", email = "rafaelbochibarros@gmail.com"},
]
dependencies = [
    "django>=5.0.4",
    "djangorestframework>=3.15.1",
    "setuptools>=69.2.0",
    "django-cors-headers>=4.3.1",
    "python-magic>=0.4.27",
    "Pillow>=10.3.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"